remove the compiler warning c4018 of microsoft visual studio.

+++ src/zxing/zxing/common/BitArray.cpp
@@ -43,7 +43,7 @@ BitArray::BitArray(int size_)
 BitArray::BitArray(std::vector<int> other)
     : size(other.size()), bits(other.size())
 {
-    for(int i=0; i<other.size(); i++)
+    for(int i=0; i<(int)other.size(); i++)
         if(other[i])
             set(i);
 }
@@ -241,7 +241,7 @@ void BitArray::ensureCapacity(int size)
     {
         ArrayRef<int> newBits = makeArray(size);
         //memcpy(bits, newBits->, bits->size());
-        for (size_t i=0; i<bits->size(); ++i) {
+        for (size_t i=0; i<(size_t)bits->size(); ++i) {
             newBits[i] = bits[i];
         }
         bits = newBits;
@@ -263,7 +263,7 @@ void BitArray::xor_(const BitArray& other)
 
 void BitArray::toBytes(int bitOffset, std::vector<byte>& array, int offset, int numBytes) const
 {
-    if(array.size() < (numBytes + offset))
+    if((int)array.size() < (numBytes + offset))
         array.resize(numBytes + offset);
 
     for (int i = 0; i < numBytes; i++) {
@@ -282,7 +282,7 @@ const std::string BitArray::toString() const
 {
     std::stringstream result;// = new StringBuilder(2 * width * height + 2);
 
-    for (size_t i = 0; i < size; i++) {
+    for (size_t i = 0; i < (size_t)size; i++) {
       if ((i & 0x07) == 0) {
         result << ' ';
       }
